// frontend/src/model/SeaEventSummary.ts
export interface SeaEventSummary {
  uuid: string;
  label: string;
  type: string;
  date: string;
  time: string;
  CrossEntity: string;
  sitrepNumber: string;
  region: string;
  shipType: string;
  imoNumber: string;
  immatNumber: string;
  lht: string;
  casualtyNumber: string;
  missingNumber: string;
  injuredNumber: string;
}